Moved theme logic to new file, renamed vars, move system theme btn to top
- Moved `system` theme button to top (since its default).
- Moved theme logic to new `theme.ts` file & renamed some variables.
- Store: Revert change that sets updateStore=true when rehydrating theme (not necessary since we update store manually)
