+++ src/lib/util/helpers.ts
@@ -1,11 +1,8 @@
-import { browser } from "$app/environment";
-import { store } from "@/store.svelte";
 import {
 	UserPermission,
 	type Icon,
 	type MediaType,
 	type TMDBContentCreditsCrew,
-	type Theme,
 	type TokenClaims,
 	type Watched,
 	type WatchedStatus,
@@ -293,55 +290,6 @@ export function getOrdinalSuffix(i: number) {
 		return "rd";
 	}
 	return "th";
-}
-
-/**
- * Utility function to handle media queries for theme preferences.
- */
-
-let mediaQuery: MediaQueryList | null = null;
-if (browser) {
-	mediaQuery = window.matchMedia("(prefers-color-scheme: dark)");
-}
-const handler = (e: MediaQueryListEvent) => {
-	document.documentElement.classList.toggle("theme-dark", e.matches);
-};
-
-/**
- * Toggle site wide theme.
- * @param theme The theme to switch to.
- * @param updateStore Should the store be updated to new theme?
- * **If set to `false`, state should be manually updated.**
- */
-export function toggleTheme(theme: Theme, updateStore = true) {
-	if (updateStore) {
-		store.appTheme = theme;
-	}
-
-	if (!mediaQuery) {
-		return;
-	}
-
-	switch (theme) {
-		case "dark":
-			document.documentElement.classList.add("theme-dark");
-			break;
-		case "light":
-			document.documentElement.classList.remove("theme-dark");
-			break;
-		case "system":
-			document.documentElement.classList.toggle(
-				"theme-dark",
-				mediaQuery.matches,
-			);
-			break;
-	}
-
-	if (theme === "dark" || theme === "light") {
-		mediaQuery.removeEventListener("change", handler);
-	} else if (theme === "system") {
-		mediaQuery.addEventListener("change", handler);
-	}
 }
 
 export function parseTokenPayload(): TokenClaims | undefined {
