+++ src/routes/(app)/profile/+page.svelte
@@ -7,11 +7,7 @@
 	import Stat from "@/lib/stats/Stat.svelte";
 	import Stats from "@/lib/stats/Stats.svelte";
 	import { updateUserSetting } from "@/lib/util/api";
-	import {
-		getOrdinalSuffix,
-		monthsShort,
-		toggleTheme,
-	} from "@/lib/util/helpers";
+	import { getOrdinalSuffix, monthsShort } from "@/lib/util/helpers";
 	import { store } from "@/store.svelte";
 	import { UserType, type Image, type Profile } from "@/types";
 	import axios from "axios";
@@ -21,6 +17,7 @@
 	import SyncModal from "./modals/SyncModal.svelte";
 	import RegionDropDown from "@/lib/RegionDropDown.svelte";
 	import RatingSetting from "@/lib/rating/RatingSetting.svelte";
+	import { toggleTheme } from "@/lib/util/theme";
 
 	let user = $derived(store.userInfo);
 	let settings = $derived(store.userSettings);
@@ -219,6 +216,13 @@
 				<h4 class="norm">Theme</h4>
 				<div class="row">
 					<button
+						class={`plain${selectedTheme === "system" ? " selected" : ""}`}
+						id="system"
+						onclick={() => toggleTheme("system")}
+					>
+						<span>system</span>
+					</button>
+					<button
 						class={`plain${selectedTheme === "light" ? " selected" : ""}`}
 						id="light"
 						onclick={() => toggleTheme("light")}
@@ -231,13 +235,6 @@
 						onclick={() => toggleTheme("dark")}
 					>
 						dark
-					</button>
-					<button
-						class={`plain${selectedTheme === "system" ? " selected" : ""}`}
-						id="system"
-						onclick={() => toggleTheme("system")}
-					>
-						<span>system</span>
 					</button>
 				</div>
 			</div>
